requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Azure AI Services
azure-ai-formrecognizer==3.3.2
azure-cognitiveservices-language-textanalytics==5.3.0
azure-storage-blob==12.19.0
openai==1.3.6

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.3

# Web interface
jinja2==3.1.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0